Moving the classes do package com.google.developers.wallet.rest Renaming methods to comply with the Java code conventions Some code improvements
