+++ java/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'wallet-samples'
+rootProject.name = 'java-rest-samples'
