+++ java/src/main/java/com/google/developers/wallet/rest/DemoEventTicket.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.google.developers.wallet.rest;
 
 // [START setup]
 // [START imports]
@@ -35,8 +36,8 @@ import java.security.interfaces.RSAPrivateKey;
 import java.util.*;
 // [END imports]
 
-/** Demo class for creating and managing Gift cards in Google Wallet. */
-public class DemoGiftCard {
+/** Demo class for creating and managing Event tickets in Google Wallet. */
+public class DemoEventTicket {
   /**
    * Path to service account key file from Google Cloud Console. Environment variable:
    * GOOGLE_APPLICATION_CREDENTIALS.
@@ -49,11 +50,11 @@ public class DemoGiftCard {
   /** Google Wallet service client. */
   public static Walletobjects service;
 
-  public DemoGiftCard() throws Exception {
+  public DemoEventTicket() throws Exception {
     keyFilePath =
         System.getenv().getOrDefault("GOOGLE_APPLICATION_CREDENTIALS", "/path/to/key.json");
 
-    Auth();
+    auth();
   }
   // [END setup]
 
@@ -61,14 +62,13 @@ public class DemoGiftCard {
   /**
    * Create authenticated HTTP client using a service account file.
    *
-   * @throws Exception
    */
-  public void Auth() throws Exception {
+  public void auth() throws Exception {
     String scope = "https://www.googleapis.com/auth/wallet_object.issuer";
 
     credentials =
         GoogleCredentials.fromStream(new FileInputStream(keyFilePath))
-            .createScoped(Arrays.asList(scope));
+            .createScoped(List.of(scope));
     credentials.refresh();
 
     HttpTransport httpTransport = GoogleNetHttpTransport.newTrustedTransport();
@@ -91,14 +91,13 @@ public class DemoGiftCard {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String CreateClass(String issuerId, String classSuffix) throws IOException {
+  public String createClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
-      service.giftcardclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.eventticketclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
 
-      System.out.println(String.format("Class %s.%s already exists!", issuerId, classSuffix));
+      System.out.printf("Class %s.%s already exists!%n", issuerId, classSuffix);
       return String.format("%s.%s", issuerId, classSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() != 404) {
@@ -109,14 +108,19 @@ public class DemoGiftCard {
     }
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardclass
-    GiftCardClass newClass =
-        new GiftCardClass()
+    // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketclass
+    EventTicketClass newClass =
+        new EventTicketClass()
+            .setEventId(String.format("%s.%s", issuerId, classSuffix))
+            .setEventName(
+                new LocalizedString()
+                    .setDefaultValue(
+                        new TranslatedString().setLanguage("en-US").setValue("Event name")))
             .setId(String.format("%s.%s", issuerId, classSuffix))
             .setIssuerName("Issuer name")
             .setReviewStatus("UNDER_REVIEW");
 
-    GiftCardClass response = service.giftcardclass().insert(newClass).execute();
+    EventTicketClass response = service.eventticketclass().insert(newClass).execute();
 
     System.out.println("Class insert response");
     System.out.println(response.toPrettyString());
@@ -134,19 +138,18 @@ public class DemoGiftCard {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String UpdateClass(String issuerId, String classSuffix) throws IOException {
-    GiftCardClass updatedClass;
+  public String updateClass(String issuerId, String classSuffix) throws IOException {
+    EventTicketClass updatedClass;
 
     // Check if the class exists
     try {
       updatedClass =
-          service.giftcardclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+          service.eventticketclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -165,9 +168,9 @@ public class DemoGiftCard {
     // Note: reviewStatus must be 'UNDER_REVIEW' or 'DRAFT' for updates
     updatedClass.setReviewStatus("UNDER_REVIEW");
 
-    GiftCardClass response =
+    EventTicketClass response =
         service
-            .giftcardclass()
+            .eventticketclass()
             .update(String.format("%s.%s", issuerId, classSuffix), updatedClass)
             .execute();
 
@@ -187,16 +190,15 @@ public class DemoGiftCard {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String PatchClass(String issuerId, String classSuffix) throws IOException {
+  public String patchClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
-      service.giftcardclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.eventticketclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -207,8 +209,8 @@ public class DemoGiftCard {
 
     // Class exists
     // Patch the class by adding a homepage
-    GiftCardClass patchBody =
-        new GiftCardClass()
+    EventTicketClass patchBody =
+        new EventTicketClass()
             .setHomepageUri(
                 new Uri()
                     .setUri("https://developers.google.com/wallet")
@@ -217,9 +219,9 @@ public class DemoGiftCard {
             // Note: reviewStatus must be 'UNDER_REVIEW' or 'DRAFT' for updates
             .setReviewStatus("UNDER_REVIEW");
 
-    GiftCardClass response =
+    EventTicketClass response =
         service
-            .giftcardclass()
+            .eventticketclass()
             .patch(String.format("%s.%s", issuerId, classSuffix), patchBody)
             .execute();
 
@@ -239,17 +241,16 @@ public class DemoGiftCard {
    * @param header The message header.
    * @param body The message body.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String AddClassMessage(String issuerId, String classSuffix, String header, String body)
+  public String addClassMessage(String issuerId, String classSuffix, String header, String body)
       throws IOException {
     // Check if the class exists
     try {
-      service.giftcardclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.eventticketclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -261,9 +262,9 @@ public class DemoGiftCard {
     AddMessageRequest message =
         new AddMessageRequest().setMessage(new Message().setHeader(header).setBody(body));
 
-    GiftCardClassAddMessageResponse response =
+    EventTicketClassAddMessageResponse response =
         service
-            .giftcardclass()
+            .eventticketclass()
             .addmessage(String.format("%s.%s", issuerId, classSuffix), message)
             .execute();
 
@@ -282,15 +283,14 @@ public class DemoGiftCard {
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String CreateObject(String issuerId, String classSuffix, String objectSuffix)
+  public String createObject(String issuerId, String classSuffix, String objectSuffix)
       throws IOException {
     // Check if the object exists
     try {
-      service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
 
-      System.out.println(String.format("Object %s.%s already exists!", issuerId, objectSuffix));
+      System.out.printf("Object %s.%s already exists!%n", issuerId, objectSuffix);
       return String.format("%s.%s", issuerId, objectSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() != 404) {
@@ -301,9 +301,9 @@ public class DemoGiftCard {
     }
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardobject
-    GiftCardObject newObject =
-        new GiftCardObject()
+    // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketobject
+    EventTicketObject newObject =
+        new EventTicketObject()
             .setId(String.format("%s.%s", issuerId, objectSuffix))
             .setClassId(String.format("%s.%s", issuerId, classSuffix))
             .setState("ACTIVE")
@@ -320,11 +320,11 @@ public class DemoGiftCard {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -338,33 +338,49 @@ public class DemoGiftCard {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
-            .setCardNumber("Card number")
-            .setPin("1234")
-            .setBalance(new Money().setMicros(20000000L).setCurrencyCode("USD"))
-            .setBalanceUpdateTime(new DateTime().setDate("2020-04-12T16:20:50.52-04:00"));
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
+            .setSeatInfo(
+                new EventSeat()
+                    .setSeat(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString().setLanguage("en-US").setValue("42")))
+                    .setRow(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString().setLanguage("en-US").setValue("G3")))
+                    .setSection(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString().setLanguage("en-US").setValue("5")))
+                    .setGate(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString().setLanguage("en-US").setValue("A"))))
+            .setTicketHolderName("Ticket holder name")
+            .setTicketNumber("Ticket number");
 
-    GiftCardObject response = service.giftcardobject().insert(newObject).execute();
+    EventTicketObject response = service.eventticketobject().insert(newObject).execute();
 
     System.out.println("Object insert response");
     System.out.println(response.toPrettyString());
@@ -382,19 +398,18 @@ public class DemoGiftCard {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String UpdateObject(String issuerId, String objectSuffix) throws IOException {
-    GiftCardObject updatedObject;
+  public String updateObject(String issuerId, String objectSuffix) throws IOException {
+    EventTicketObject updatedObject;
 
     // Check if the object exists
     try {
       updatedObject =
-          service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+          service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -412,14 +427,14 @@ public class DemoGiftCard {
 
     if (updatedObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
-      updatedObject.setLinksModuleData(new LinksModuleData().setUris(Arrays.asList(newLink)));
+      updatedObject.setLinksModuleData(new LinksModuleData().setUris(List.of(newLink)));
     } else {
       updatedObject.getLinksModuleData().getUris().add(newLink);
     }
 
-    GiftCardObject response =
+    EventTicketObject response =
         service
-            .giftcardobject()
+            .eventticketobject()
             .update(String.format("%s.%s", issuerId, objectSuffix), updatedObject)
             .execute();
 
@@ -437,19 +452,18 @@ public class DemoGiftCard {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String PatchObject(String issuerId, String objectSuffix) throws IOException {
-    GiftCardObject existingObject;
+  public String patchObject(String issuerId, String objectSuffix) throws IOException {
+    EventTicketObject existingObject;
 
     // Check if the object exists
     try {
       existingObject =
-          service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+          service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -465,7 +479,7 @@ public class DemoGiftCard {
             .setUri("https://developers.google.com/wallet")
             .setDescription("New link description");
 
-    GiftCardObject patchBody = new GiftCardObject();
+    EventTicketObject patchBody = new EventTicketObject();
 
     if (existingObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
@@ -475,9 +489,9 @@ public class DemoGiftCard {
     }
     patchBody.getLinksModuleData().getUris().add(newLink);
 
-    GiftCardObject response =
+    EventTicketObject response =
         service
-            .giftcardobject()
+            .eventticketobject()
             .patch(String.format("%s.%s", issuerId, objectSuffix), patchBody)
             .execute();
 
@@ -498,16 +512,15 @@ public class DemoGiftCard {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String ExpireObject(String issuerId, String objectSuffix) throws IOException {
+  public String expireObject(String issuerId, String objectSuffix) throws IOException {
     // Check if the object exists
     try {
-      service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -517,11 +530,11 @@ public class DemoGiftCard {
     }
 
     // Patch the object, setting the pass as expired
-    GiftCardObject patchBody = new GiftCardObject().setState("EXPIRED");
+    EventTicketObject patchBody = new EventTicketObject().setState("EXPIRED");
 
-    GiftCardObject response =
+    EventTicketObject response =
         service
-            .giftcardobject()
+            .eventticketobject()
             .patch(String.format("%s.%s", issuerId, objectSuffix), patchBody)
             .execute();
 
@@ -541,17 +554,16 @@ public class DemoGiftCard {
    * @param header The message header.
    * @param body The message body.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String AddObjectMessage(String issuerId, String objectSuffix, String header, String body)
+  public String addObjectMessage(String issuerId, String objectSuffix, String header, String body)
       throws IOException {
     // Check if the object exists
     try {
-      service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -563,9 +575,9 @@ public class DemoGiftCard {
     AddMessageRequest message =
         new AddMessageRequest().setMessage(new Message().setHeader(header).setBody(body));
 
-    GiftCardObjectAddMessageResponse response =
+    EventTicketObjectAddMessageResponse response =
         service
-            .giftcardobject()
+            .eventticketobject()
             .addmessage(String.format("%s.%s", issuerId, objectSuffix), message)
             .execute();
 
@@ -589,19 +601,23 @@ public class DemoGiftCard {
    * @param objectSuffix Developer-defined unique ID for the pass object.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
+  public String createJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardclass
-    GiftCardClass newClass =
-        new GiftCardClass()
+    // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketclass
+    EventTicketClass newClass =
+        new EventTicketClass()
             .setId(String.format("%s.%s", issuerId, classSuffix))
             .setIssuerName("Issuer name")
-            .setReviewStatus("UNDER_REVIEW");
+            .setReviewStatus("UNDER_REVIEW")
+            .setEventName(
+                new LocalizedString()
+                    .setDefaultValue(
+                        new TranslatedString().setLanguage("en-US").setValue("Event name")));
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardobject
-    GiftCardObject newObject =
-        new GiftCardObject()
+    // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketobject
+    EventTicketObject newObject =
+        new EventTicketObject()
             .setId(String.format("%s.%s", issuerId, objectSuffix))
             .setClassId(String.format("%s.%s", issuerId, classSuffix))
             .setState("ACTIVE")
@@ -618,11 +634,11 @@ public class DemoGiftCard {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -636,43 +652,59 @@ public class DemoGiftCard {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
-            .setCardNumber("Card number")
-            .setPin("1234")
-            .setBalance(new Money().setMicros(20000000L).setCurrencyCode("USD"))
-            .setBalanceUpdateTime(new DateTime().setDate("2020-04-12T16:20:50.52-04:00"));
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
+            .setSeatInfo(
+                new EventSeat()
+                    .setSeat(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString().setLanguage("en-US").setValue("42")))
+                    .setRow(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString().setLanguage("en-US").setValue("G3")))
+                    .setSection(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString().setLanguage("en-US").setValue("5")))
+                    .setGate(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString().setLanguage("en-US").setValue("A"))))
+            .setTicketHolderName("Ticket holder name")
+            .setTicketNumber("Ticket number");
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
 
     // Create the Google Wallet payload and add to the JWT
     HashMap<String, Object> payload = new HashMap<String, Object>();
-    payload.put("giftCardClasses", Arrays.asList(newClass));
-    payload.put("giftCardObjects", Arrays.asList(newObject));
+    payload.put("eventTicketClasses", List.of(newClass));
+    payload.put("eventTicketObjects", List.of(newObject));
     claims.put("payload", payload);
 
     // The service account credentials are used to sign the JWT
@@ -682,7 +714,7 @@ public class DemoGiftCard {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -703,7 +735,7 @@ public class DemoGiftCard {
    * @param issuerId The issuer ID being used for this request.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTExistingObjects(String issuerId) {
+  public String createJWTExistingObjects(String issuerId) {
     // Multiple pass types can be added at the same time
     // At least one type must be specified in the JWT claims
     // Note: Make sure to replace the placeholder class and object suffixes
@@ -712,64 +744,64 @@ public class DemoGiftCard {
     // Event tickets
     objectsToAdd.put(
         "eventTicketObjects",
-        Arrays.asList(
-            new EventTicketObject()
-                .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
+            List.of(
+                    new EventTicketObject()
+                            .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
 
     // Boarding passes
     objectsToAdd.put(
         "flightObjects",
-        Arrays.asList(
-            new FlightObject()
-                .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
+            List.of(
+                    new FlightObject()
+                            .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
 
     // Generic passes
     objectsToAdd.put(
         "genericObjects",
-        Arrays.asList(
-            new GenericObject()
-                .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
+            List.of(
+                    new GenericObject()
+                            .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
 
     // Gift cards
     objectsToAdd.put(
         "giftCardObjects",
-        Arrays.asList(
-            new GiftCardObject()
-                .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
+            List.of(
+                    new GiftCardObject()
+                            .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
 
     // Loyalty cards
     objectsToAdd.put(
         "loyaltyObjects",
-        Arrays.asList(
-            new LoyaltyObject()
-                .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
+            List.of(
+                    new LoyaltyObject()
+                            .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
 
     // Offers
     objectsToAdd.put(
         "offerObjects",
-        Arrays.asList(
-            new OfferObject()
-                .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
+            List.of(
+                    new OfferObject()
+                            .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
 
     // Transit passes
     objectsToAdd.put(
         "transitObjects",
-        Arrays.asList(
-            new TransitObject()
-                .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
+            List.of(
+                    new TransitObject()
+                            .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
     claims.put("payload", objectsToAdd);
 
@@ -780,7 +812,7 @@ public class DemoGiftCard {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -792,17 +824,16 @@ public class DemoGiftCard {
    *
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
-   * @throws IOException
    */
-  public void BatchCreateObjects(String issuerId, String classSuffix) throws IOException {
+  public void batchCreateObjects(String issuerId, String classSuffix) throws IOException {
     // Create the batch request client
     BatchRequest batch = service.batch(new HttpCredentialsAdapter(credentials));
 
     // The callback will be invoked for each request in the batch
-    JsonBatchCallback<GiftCardObject> callback =
-        new JsonBatchCallback<GiftCardObject>() {
+    JsonBatchCallback<EventTicketObject> callback =
+        new JsonBatchCallback<EventTicketObject>() {
           // Invoked if the request was successful
-          public void onSuccess(GiftCardObject response, HttpHeaders responseHeaders) {
+          public void onSuccess(EventTicketObject response, HttpHeaders responseHeaders) {
             System.out.println("Batch insert response");
             System.out.println(response.toString());
           }
@@ -819,9 +850,9 @@ public class DemoGiftCard {
       String objectSuffix = UUID.randomUUID().toString().replaceAll("[^\\w.-]", "_");
 
       // See link below for more information on required properties
-      // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardobject
-      GiftCardObject batchObject =
-          new GiftCardObject()
+      // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketobject
+      EventTicketObject batchObject =
+          new EventTicketObject()
               .setId(String.format("%s.%s", issuerId, objectSuffix))
               .setClassId(String.format("%s.%s", issuerId, classSuffix))
               .setState("ACTIVE")
@@ -838,11 +869,11 @@ public class DemoGiftCard {
                                       .setLanguage("en-US")
                                       .setValue("Hero image description"))))
               .setTextModulesData(
-                  Arrays.asList(
-                      new TextModuleData()
-                          .setHeader("Text module header")
-                          .setBody("Text module body")
-                          .setId("TEXT_MODULE_ID")))
+                      List.of(
+                              new TextModuleData()
+                                      .setHeader("Text module header")
+                                      .setBody("Text module body")
+                                      .setId("TEXT_MODULE_ID")))
               .setLinksModuleData(
                   new LinksModuleData()
                       .setUris(
@@ -856,33 +887,49 @@ public class DemoGiftCard {
                                   .setDescription("Link module tel description")
                                   .setId("LINK_MODULE_TEL_ID"))))
               .setImageModulesData(
-                  Arrays.asList(
-                      new ImageModuleData()
-                          .setMainImage(
-                              new Image()
-                                  .setSourceUri(
-                                      new ImageUri()
-                                          .setUri(
-                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                  .setContentDescription(
-                                      new LocalizedString()
-                                          .setDefaultValue(
-                                              new TranslatedString()
-                                                  .setLanguage("en-US")
-                                                  .setValue("Image module description"))))
-                          .setId("IMAGE_MODULE_ID")))
+                      List.of(
+                              new ImageModuleData()
+                                      .setMainImage(
+                                              new Image()
+                                                      .setSourceUri(
+                                                              new ImageUri()
+                                                                      .setUri(
+                                                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                      .setContentDescription(
+                                                              new LocalizedString()
+                                                                      .setDefaultValue(
+                                                                              new TranslatedString()
+                                                                                      .setLanguage("en-US")
+                                                                                      .setValue("Image module description"))))
+                                      .setId("IMAGE_MODULE_ID")))
               .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
               .setLocations(
-                  Arrays.asList(
-                      new LatLongPoint()
-                          .setLatitude(37.424015499999996)
-                          .setLongitude(-122.09259560000001)))
-              .setCardNumber("Card number")
-              .setPin("1234")
-              .setBalance(new Money().setMicros(20000000L).setCurrencyCode("USD"))
-              .setBalanceUpdateTime(new DateTime().setDate("2020-04-12T16:20:50.52-04:00"));
+                      List.of(
+                              new LatLongPoint()
+                                      .setLatitude(37.424015499999996)
+                                      .setLongitude(-122.09259560000001)))
+              .setSeatInfo(
+                  new EventSeat()
+                      .setSeat(
+                          new LocalizedString()
+                              .setDefaultValue(
+                                  new TranslatedString().setLanguage("en-US").setValue("42")))
+                      .setRow(
+                          new LocalizedString()
+                              .setDefaultValue(
+                                  new TranslatedString().setLanguage("en-US").setValue("G3")))
+                      .setSection(
+                          new LocalizedString()
+                              .setDefaultValue(
+                                  new TranslatedString().setLanguage("en-US").setValue("5")))
+                      .setGate(
+                          new LocalizedString()
+                              .setDefaultValue(
+                                  new TranslatedString().setLanguage("en-US").setValue("A"))))
+              .setTicketHolderName("Ticket holder name")
+              .setTicketNumber("Ticket number");
 
-      service.giftcardobject().insert(batchObject).queue(batch, callback);
+      service.eventticketobject().insert(batchObject).queue(batch, callback);
     }
 
     // Invoke the batch API calls
