+++ java/src/main/java/com/google/developers/wallet/rest/DemoGiftCard.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.google.developers.wallet.rest;
 
 // [START setup]
 // [START imports]
@@ -35,8 +36,8 @@ import java.security.interfaces.RSAPrivateKey;
 import java.util.*;
 // [END imports]
 
-/** Demo class for creating and managing Event tickets in Google Wallet. */
-public class DemoEventTicket {
+/** Demo class for creating and managing Gift cards in Google Wallet. */
+public class DemoGiftCard {
   /**
    * Path to service account key file from Google Cloud Console. Environment variable:
    * GOOGLE_APPLICATION_CREDENTIALS.
@@ -49,11 +50,11 @@ public class DemoEventTicket {
   /** Google Wallet service client. */
   public static Walletobjects service;
 
-  public DemoEventTicket() throws Exception {
+  public DemoGiftCard() throws Exception {
     keyFilePath =
         System.getenv().getOrDefault("GOOGLE_APPLICATION_CREDENTIALS", "/path/to/key.json");
 
-    Auth();
+    auth();
   }
   // [END setup]
 
@@ -61,14 +62,13 @@ public class DemoEventTicket {
   /**
    * Create authenticated HTTP client using a service account file.
    *
-   * @throws Exception
    */
-  public void Auth() throws Exception {
+  public void auth() throws Exception {
     String scope = "https://www.googleapis.com/auth/wallet_object.issuer";
 
     credentials =
         GoogleCredentials.fromStream(new FileInputStream(keyFilePath))
-            .createScoped(Arrays.asList(scope));
+            .createScoped(List.of(scope));
     credentials.refresh();
 
     HttpTransport httpTransport = GoogleNetHttpTransport.newTrustedTransport();
@@ -91,14 +91,13 @@ public class DemoEventTicket {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String CreateClass(String issuerId, String classSuffix) throws IOException {
+  public String createClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
-      service.eventticketclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.giftcardclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
 
-      System.out.println(String.format("Class %s.%s already exists!", issuerId, classSuffix));
+      System.out.printf("Class %s.%s already exists!%n", issuerId, classSuffix);
       return String.format("%s.%s", issuerId, classSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() != 404) {
@@ -109,19 +108,14 @@ public class DemoEventTicket {
     }
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketclass
-    EventTicketClass newClass =
-        new EventTicketClass()
-            .setEventId(String.format("%s.%s", issuerId, classSuffix))
-            .setEventName(
-                new LocalizedString()
-                    .setDefaultValue(
-                        new TranslatedString().setLanguage("en-US").setValue("Event name")))
+    // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardclass
+    GiftCardClass newClass =
+        new GiftCardClass()
             .setId(String.format("%s.%s", issuerId, classSuffix))
             .setIssuerName("Issuer name")
             .setReviewStatus("UNDER_REVIEW");
 
-    EventTicketClass response = service.eventticketclass().insert(newClass).execute();
+    GiftCardClass response = service.giftcardclass().insert(newClass).execute();
 
     System.out.println("Class insert response");
     System.out.println(response.toPrettyString());
@@ -139,19 +133,18 @@ public class DemoEventTicket {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String UpdateClass(String issuerId, String classSuffix) throws IOException {
-    EventTicketClass updatedClass;
+  public String updateClass(String issuerId, String classSuffix) throws IOException {
+    GiftCardClass updatedClass;
 
     // Check if the class exists
     try {
       updatedClass =
-          service.eventticketclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+          service.giftcardclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -170,9 +163,9 @@ public class DemoEventTicket {
     // Note: reviewStatus must be 'UNDER_REVIEW' or 'DRAFT' for updates
     updatedClass.setReviewStatus("UNDER_REVIEW");
 
-    EventTicketClass response =
+    GiftCardClass response =
         service
-            .eventticketclass()
+            .giftcardclass()
             .update(String.format("%s.%s", issuerId, classSuffix), updatedClass)
             .execute();
 
@@ -192,16 +185,15 @@ public class DemoEventTicket {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String PatchClass(String issuerId, String classSuffix) throws IOException {
+  public String patchClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
-      service.eventticketclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.giftcardclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -212,8 +204,8 @@ public class DemoEventTicket {
 
     // Class exists
     // Patch the class by adding a homepage
-    EventTicketClass patchBody =
-        new EventTicketClass()
+    GiftCardClass patchBody =
+        new GiftCardClass()
             .setHomepageUri(
                 new Uri()
                     .setUri("https://developers.google.com/wallet")
@@ -222,9 +214,9 @@ public class DemoEventTicket {
             // Note: reviewStatus must be 'UNDER_REVIEW' or 'DRAFT' for updates
             .setReviewStatus("UNDER_REVIEW");
 
-    EventTicketClass response =
+    GiftCardClass response =
         service
-            .eventticketclass()
+            .giftcardclass()
             .patch(String.format("%s.%s", issuerId, classSuffix), patchBody)
             .execute();
 
@@ -244,17 +236,16 @@ public class DemoEventTicket {
    * @param header The message header.
    * @param body The message body.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String AddClassMessage(String issuerId, String classSuffix, String header, String body)
+  public String addClassMessage(String issuerId, String classSuffix, String header, String body)
       throws IOException {
     // Check if the class exists
     try {
-      service.eventticketclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.giftcardclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -266,9 +257,9 @@ public class DemoEventTicket {
     AddMessageRequest message =
         new AddMessageRequest().setMessage(new Message().setHeader(header).setBody(body));
 
-    EventTicketClassAddMessageResponse response =
+    GiftCardClassAddMessageResponse response =
         service
-            .eventticketclass()
+            .giftcardclass()
             .addmessage(String.format("%s.%s", issuerId, classSuffix), message)
             .execute();
 
@@ -287,15 +278,14 @@ public class DemoEventTicket {
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String CreateObject(String issuerId, String classSuffix, String objectSuffix)
+  public String createObject(String issuerId, String classSuffix, String objectSuffix)
       throws IOException {
     // Check if the object exists
     try {
-      service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
 
-      System.out.println(String.format("Object %s.%s already exists!", issuerId, objectSuffix));
+      System.out.printf("Object %s.%s already exists!%n", issuerId, objectSuffix);
       return String.format("%s.%s", issuerId, objectSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() != 404) {
@@ -306,9 +296,9 @@ public class DemoEventTicket {
     }
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketobject
-    EventTicketObject newObject =
-        new EventTicketObject()
+    // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardobject
+    GiftCardObject newObject =
+        new GiftCardObject()
             .setId(String.format("%s.%s", issuerId, objectSuffix))
             .setClassId(String.format("%s.%s", issuerId, classSuffix))
             .setState("ACTIVE")
@@ -325,11 +315,11 @@ public class DemoEventTicket {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -343,49 +333,33 @@ public class DemoEventTicket {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
-            .setSeatInfo(
-                new EventSeat()
-                    .setSeat(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString().setLanguage("en-US").setValue("42")))
-                    .setRow(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString().setLanguage("en-US").setValue("G3")))
-                    .setSection(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString().setLanguage("en-US").setValue("5")))
-                    .setGate(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString().setLanguage("en-US").setValue("A"))))
-            .setTicketHolderName("Ticket holder name")
-            .setTicketNumber("Ticket number");
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
+            .setCardNumber("Card number")
+            .setPin("1234")
+            .setBalance(new Money().setMicros(20000000L).setCurrencyCode("USD"))
+            .setBalanceUpdateTime(new DateTime().setDate("2020-04-12T16:20:50.52-04:00"));
 
-    EventTicketObject response = service.eventticketobject().insert(newObject).execute();
+    GiftCardObject response = service.giftcardobject().insert(newObject).execute();
 
     System.out.println("Object insert response");
     System.out.println(response.toPrettyString());
@@ -403,19 +377,18 @@ public class DemoEventTicket {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String UpdateObject(String issuerId, String objectSuffix) throws IOException {
-    EventTicketObject updatedObject;
+  public String updateObject(String issuerId, String objectSuffix) throws IOException {
+    GiftCardObject updatedObject;
 
     // Check if the object exists
     try {
       updatedObject =
-          service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+          service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -433,14 +406,14 @@ public class DemoEventTicket {
 
     if (updatedObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
-      updatedObject.setLinksModuleData(new LinksModuleData().setUris(Arrays.asList(newLink)));
+      updatedObject.setLinksModuleData(new LinksModuleData().setUris(List.of(newLink)));
     } else {
       updatedObject.getLinksModuleData().getUris().add(newLink);
     }
 
-    EventTicketObject response =
+    GiftCardObject response =
         service
-            .eventticketobject()
+            .giftcardobject()
             .update(String.format("%s.%s", issuerId, objectSuffix), updatedObject)
             .execute();
 
@@ -458,19 +431,18 @@ public class DemoEventTicket {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String PatchObject(String issuerId, String objectSuffix) throws IOException {
-    EventTicketObject existingObject;
+  public String patchObject(String issuerId, String objectSuffix) throws IOException {
+    GiftCardObject existingObject;
 
     // Check if the object exists
     try {
       existingObject =
-          service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+          service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -486,7 +458,7 @@ public class DemoEventTicket {
             .setUri("https://developers.google.com/wallet")
             .setDescription("New link description");
 
-    EventTicketObject patchBody = new EventTicketObject();
+    GiftCardObject patchBody = new GiftCardObject();
 
     if (existingObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
@@ -496,9 +468,9 @@ public class DemoEventTicket {
     }
     patchBody.getLinksModuleData().getUris().add(newLink);
 
-    EventTicketObject response =
+    GiftCardObject response =
         service
-            .eventticketobject()
+            .giftcardobject()
             .patch(String.format("%s.%s", issuerId, objectSuffix), patchBody)
             .execute();
 
@@ -519,16 +491,15 @@ public class DemoEventTicket {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String ExpireObject(String issuerId, String objectSuffix) throws IOException {
+  public String expireObject(String issuerId, String objectSuffix) throws IOException {
     // Check if the object exists
     try {
-      service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -538,11 +509,11 @@ public class DemoEventTicket {
     }
 
     // Patch the object, setting the pass as expired
-    EventTicketObject patchBody = new EventTicketObject().setState("EXPIRED");
+    GiftCardObject patchBody = new GiftCardObject().setState("EXPIRED");
 
-    EventTicketObject response =
+    GiftCardObject response =
         service
-            .eventticketobject()
+            .giftcardobject()
             .patch(String.format("%s.%s", issuerId, objectSuffix), patchBody)
             .execute();
 
@@ -562,17 +533,16 @@ public class DemoEventTicket {
    * @param header The message header.
    * @param body The message body.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String AddObjectMessage(String issuerId, String objectSuffix, String header, String body)
+  public String addObjectMessage(String issuerId, String objectSuffix, String header, String body)
       throws IOException {
     // Check if the object exists
     try {
-      service.eventticketobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.giftcardobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -584,9 +554,9 @@ public class DemoEventTicket {
     AddMessageRequest message =
         new AddMessageRequest().setMessage(new Message().setHeader(header).setBody(body));
 
-    EventTicketObjectAddMessageResponse response =
+    GiftCardObjectAddMessageResponse response =
         service
-            .eventticketobject()
+            .giftcardobject()
             .addmessage(String.format("%s.%s", issuerId, objectSuffix), message)
             .execute();
 
@@ -610,23 +580,19 @@ public class DemoEventTicket {
    * @param objectSuffix Developer-defined unique ID for the pass object.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
+  public String createJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketclass
-    EventTicketClass newClass =
-        new EventTicketClass()
+    // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardclass
+    GiftCardClass newClass =
+        new GiftCardClass()
             .setId(String.format("%s.%s", issuerId, classSuffix))
             .setIssuerName("Issuer name")
-            .setReviewStatus("UNDER_REVIEW")
-            .setEventName(
-                new LocalizedString()
-                    .setDefaultValue(
-                        new TranslatedString().setLanguage("en-US").setValue("Event name")));
+            .setReviewStatus("UNDER_REVIEW");
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketobject
-    EventTicketObject newObject =
-        new EventTicketObject()
+    // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardobject
+    GiftCardObject newObject =
+        new GiftCardObject()
             .setId(String.format("%s.%s", issuerId, objectSuffix))
             .setClassId(String.format("%s.%s", issuerId, classSuffix))
             .setState("ACTIVE")
@@ -643,11 +609,11 @@ public class DemoEventTicket {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -661,59 +627,43 @@ public class DemoEventTicket {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
-            .setSeatInfo(
-                new EventSeat()
-                    .setSeat(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString().setLanguage("en-US").setValue("42")))
-                    .setRow(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString().setLanguage("en-US").setValue("G3")))
-                    .setSection(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString().setLanguage("en-US").setValue("5")))
-                    .setGate(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString().setLanguage("en-US").setValue("A"))))
-            .setTicketHolderName("Ticket holder name")
-            .setTicketNumber("Ticket number");
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
+            .setCardNumber("Card number")
+            .setPin("1234")
+            .setBalance(new Money().setMicros(20000000L).setCurrencyCode("USD"))
+            .setBalanceUpdateTime(new DateTime().setDate("2020-04-12T16:20:50.52-04:00"));
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
 
     // Create the Google Wallet payload and add to the JWT
     HashMap<String, Object> payload = new HashMap<String, Object>();
-    payload.put("eventTicketClasses", Arrays.asList(newClass));
-    payload.put("eventTicketObjects", Arrays.asList(newObject));
+    payload.put("giftCardClasses", List.of(newClass));
+    payload.put("giftCardObjects", List.of(newObject));
     claims.put("payload", payload);
 
     // The service account credentials are used to sign the JWT
@@ -723,7 +673,7 @@ public class DemoEventTicket {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -744,7 +694,7 @@ public class DemoEventTicket {
    * @param issuerId The issuer ID being used for this request.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTExistingObjects(String issuerId) {
+  public String createJWTExistingObjects(String issuerId) {
     // Multiple pass types can be added at the same time
     // At least one type must be specified in the JWT claims
     // Note: Make sure to replace the placeholder class and object suffixes
@@ -753,64 +703,64 @@ public class DemoEventTicket {
     // Event tickets
     objectsToAdd.put(
         "eventTicketObjects",
-        Arrays.asList(
-            new EventTicketObject()
-                .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
+            List.of(
+                    new EventTicketObject()
+                            .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
 
     // Boarding passes
     objectsToAdd.put(
         "flightObjects",
-        Arrays.asList(
-            new FlightObject()
-                .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
+            List.of(
+                    new FlightObject()
+                            .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
 
     // Generic passes
     objectsToAdd.put(
         "genericObjects",
-        Arrays.asList(
-            new GenericObject()
-                .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
+            List.of(
+                    new GenericObject()
+                            .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
 
     // Gift cards
     objectsToAdd.put(
         "giftCardObjects",
-        Arrays.asList(
-            new GiftCardObject()
-                .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
+            List.of(
+                    new GiftCardObject()
+                            .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
 
     // Loyalty cards
     objectsToAdd.put(
         "loyaltyObjects",
-        Arrays.asList(
-            new LoyaltyObject()
-                .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
+            List.of(
+                    new LoyaltyObject()
+                            .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
 
     // Offers
     objectsToAdd.put(
         "offerObjects",
-        Arrays.asList(
-            new OfferObject()
-                .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
+            List.of(
+                    new OfferObject()
+                            .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
 
     // Transit passes
     objectsToAdd.put(
         "transitObjects",
-        Arrays.asList(
-            new TransitObject()
-                .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
+            List.of(
+                    new TransitObject()
+                            .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
     claims.put("payload", objectsToAdd);
 
@@ -821,7 +771,7 @@ public class DemoEventTicket {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -833,17 +783,16 @@ public class DemoEventTicket {
    *
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
-   * @throws IOException
    */
-  public void BatchCreateObjects(String issuerId, String classSuffix) throws IOException {
+  public void batchCreateObjects(String issuerId, String classSuffix) throws IOException {
     // Create the batch request client
     BatchRequest batch = service.batch(new HttpCredentialsAdapter(credentials));
 
     // The callback will be invoked for each request in the batch
-    JsonBatchCallback<EventTicketObject> callback =
-        new JsonBatchCallback<EventTicketObject>() {
+    JsonBatchCallback<GiftCardObject> callback =
+        new JsonBatchCallback<GiftCardObject>() {
           // Invoked if the request was successful
-          public void onSuccess(EventTicketObject response, HttpHeaders responseHeaders) {
+          public void onSuccess(GiftCardObject response, HttpHeaders responseHeaders) {
             System.out.println("Batch insert response");
             System.out.println(response.toString());
           }
@@ -860,9 +809,9 @@ public class DemoEventTicket {
       String objectSuffix = UUID.randomUUID().toString().replaceAll("[^\\w.-]", "_");
 
       // See link below for more information on required properties
-      // https://developers.google.com/wallet/tickets/events/rest/v1/eventticketobject
-      EventTicketObject batchObject =
-          new EventTicketObject()
+      // https://developers.google.com/wallet/retail/gift-cards/rest/v1/giftcardobject
+      GiftCardObject batchObject =
+          new GiftCardObject()
               .setId(String.format("%s.%s", issuerId, objectSuffix))
               .setClassId(String.format("%s.%s", issuerId, classSuffix))
               .setState("ACTIVE")
@@ -879,11 +828,11 @@ public class DemoEventTicket {
                                       .setLanguage("en-US")
                                       .setValue("Hero image description"))))
               .setTextModulesData(
-                  Arrays.asList(
-                      new TextModuleData()
-                          .setHeader("Text module header")
-                          .setBody("Text module body")
-                          .setId("TEXT_MODULE_ID")))
+                      List.of(
+                              new TextModuleData()
+                                      .setHeader("Text module header")
+                                      .setBody("Text module body")
+                                      .setId("TEXT_MODULE_ID")))
               .setLinksModuleData(
                   new LinksModuleData()
                       .setUris(
@@ -897,49 +846,33 @@ public class DemoEventTicket {
                                   .setDescription("Link module tel description")
                                   .setId("LINK_MODULE_TEL_ID"))))
               .setImageModulesData(
-                  Arrays.asList(
-                      new ImageModuleData()
-                          .setMainImage(
-                              new Image()
-                                  .setSourceUri(
-                                      new ImageUri()
-                                          .setUri(
-                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                  .setContentDescription(
-                                      new LocalizedString()
-                                          .setDefaultValue(
-                                              new TranslatedString()
-                                                  .setLanguage("en-US")
-                                                  .setValue("Image module description"))))
-                          .setId("IMAGE_MODULE_ID")))
+                      List.of(
+                              new ImageModuleData()
+                                      .setMainImage(
+                                              new Image()
+                                                      .setSourceUri(
+                                                              new ImageUri()
+                                                                      .setUri(
+                                                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                      .setContentDescription(
+                                                              new LocalizedString()
+                                                                      .setDefaultValue(
+                                                                              new TranslatedString()
+                                                                                      .setLanguage("en-US")
+                                                                                      .setValue("Image module description"))))
+                                      .setId("IMAGE_MODULE_ID")))
               .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
               .setLocations(
-                  Arrays.asList(
-                      new LatLongPoint()
-                          .setLatitude(37.424015499999996)
-                          .setLongitude(-122.09259560000001)))
-              .setSeatInfo(
-                  new EventSeat()
-                      .setSeat(
-                          new LocalizedString()
-                              .setDefaultValue(
-                                  new TranslatedString().setLanguage("en-US").setValue("42")))
-                      .setRow(
-                          new LocalizedString()
-                              .setDefaultValue(
-                                  new TranslatedString().setLanguage("en-US").setValue("G3")))
-                      .setSection(
-                          new LocalizedString()
-                              .setDefaultValue(
-                                  new TranslatedString().setLanguage("en-US").setValue("5")))
-                      .setGate(
-                          new LocalizedString()
-                              .setDefaultValue(
-                                  new TranslatedString().setLanguage("en-US").setValue("A"))))
-              .setTicketHolderName("Ticket holder name")
-              .setTicketNumber("Ticket number");
+                      List.of(
+                              new LatLongPoint()
+                                      .setLatitude(37.424015499999996)
+                                      .setLongitude(-122.09259560000001)))
+              .setCardNumber("Card number")
+              .setPin("1234")
+              .setBalance(new Money().setMicros(20000000L).setCurrencyCode("USD"))
+              .setBalanceUpdateTime(new DateTime().setDate("2020-04-12T16:20:50.52-04:00"));
 
-      service.eventticketobject().insert(batchObject).queue(batch, callback);
+      service.giftcardobject().insert(batchObject).queue(batch, callback);
     }
 
     // Invoke the batch API calls
