+++ java/src/main/java/com/google/developers/wallet/rest/DemoLoyalty.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.google.developers.wallet.rest;
 
 // [START setup]
 // [START imports]
@@ -35,8 +36,8 @@ import java.security.interfaces.RSAPrivateKey;
 import java.util.*;
 // [END imports]
 
-/** Demo class for creating and managing Flights in Google Wallet. */
-public class DemoFlight {
+/** Demo class for creating and managing Loyalty cards in Google Wallet. */
+public class DemoLoyalty {
   /**
    * Path to service account key file from Google Cloud Console. Environment variable:
    * GOOGLE_APPLICATION_CREDENTIALS.
@@ -49,11 +50,11 @@ public class DemoFlight {
   /** Google Wallet service client. */
   public static Walletobjects service;
 
-  public DemoFlight() throws Exception {
+  public DemoLoyalty() throws Exception {
     keyFilePath =
         System.getenv().getOrDefault("GOOGLE_APPLICATION_CREDENTIALS", "/path/to/key.json");
 
-    Auth();
+    auth();
   }
   // [END setup]
 
@@ -61,14 +62,13 @@ public class DemoFlight {
   /**
    * Create authenticated HTTP client using a service account file.
    *
-   * @throws Exception
    */
-  public void Auth() throws Exception {
+  public void auth() throws Exception {
     String scope = "https://www.googleapis.com/auth/wallet_object.issuer";
 
     credentials =
         GoogleCredentials.fromStream(new FileInputStream(keyFilePath))
-            .createScoped(Arrays.asList(scope));
+            .createScoped(List.of(scope));
     credentials.refresh();
 
     HttpTransport httpTransport = GoogleNetHttpTransport.newTrustedTransport();
@@ -91,14 +91,13 @@ public class DemoFlight {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String CreateClass(String issuerId, String classSuffix) throws IOException {
+  public String createClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
-      service.flightclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.loyaltyclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
 
-      System.out.println(String.format("Class %s.%s already exists!", issuerId, classSuffix));
+      System.out.printf("Class %s.%s already exists!%n", issuerId, classSuffix);
       return String.format("%s.%s", issuerId, classSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() != 404) {
@@ -109,22 +108,27 @@ public class DemoFlight {
     }
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/tickets/boarding-passes/rest/v1/flightclass
-    FlightClass newClass =
-        new FlightClass()
+    // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyclass
+    LoyaltyClass newClass =
+        new LoyaltyClass()
             .setId(String.format("%s.%s", issuerId, classSuffix))
             .setIssuerName("Issuer name")
             .setReviewStatus("UNDER_REVIEW")
-            .setLocalScheduledDepartureDateTime("2023-07-02T15:30:00")
-            .setFlightHeader(
-                new FlightHeader()
-                    .setCarrier(new FlightCarrier().setCarrierIataCode("LX"))
-                    .setFlightNumber("123"))
-            .setOrigin(new AirportInfo().setAirportIataCode("LAX").setTerminal("1").setGate("A2"))
-            .setDestination(
-                new AirportInfo().setAirportIataCode("SFO").setTerminal("2").setGate("C3"));
+            .setProgramName("Program name")
+            .setProgramLogo(
+                new Image()
+                    .setSourceUri(
+                        new ImageUri()
+                            .setUri(
+                                "https://farm4.staticflickr.com/3723/11177041115_6e6a3b6f49_o.jpg"))
+                    .setContentDescription(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString()
+                                    .setLanguage("en-US")
+                                    .setValue("Logo description"))));
 
-    FlightClass response = service.flightclass().insert(newClass).execute();
+    LoyaltyClass response = service.loyaltyclass().insert(newClass).execute();
 
     System.out.println("Class insert response");
     System.out.println(response.toPrettyString());
@@ -142,19 +146,18 @@ public class DemoFlight {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String UpdateClass(String issuerId, String classSuffix) throws IOException {
-    FlightClass updatedClass;
+  public String updateClass(String issuerId, String classSuffix) throws IOException {
+    LoyaltyClass updatedClass;
 
     // Check if the class exists
     try {
       updatedClass =
-          service.flightclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+          service.loyaltyclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -173,9 +176,9 @@ public class DemoFlight {
     // Note: reviewStatus must be 'UNDER_REVIEW' or 'DRAFT' for updates
     updatedClass.setReviewStatus("UNDER_REVIEW");
 
-    FlightClass response =
+    LoyaltyClass response =
         service
-            .flightclass()
+            .loyaltyclass()
             .update(String.format("%s.%s", issuerId, classSuffix), updatedClass)
             .execute();
 
@@ -195,16 +198,15 @@ public class DemoFlight {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String PatchClass(String issuerId, String classSuffix) throws IOException {
+  public String patchClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
-      service.flightclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.loyaltyclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -215,8 +217,8 @@ public class DemoFlight {
 
     // Class exists
     // Patch the class by adding a homepage
-    FlightClass patchBody =
-        new FlightClass()
+    LoyaltyClass patchBody =
+        new LoyaltyClass()
             .setHomepageUri(
                 new Uri()
                     .setUri("https://developers.google.com/wallet")
@@ -225,9 +227,9 @@ public class DemoFlight {
             // Note: reviewStatus must be 'UNDER_REVIEW' or 'DRAFT' for updates
             .setReviewStatus("UNDER_REVIEW");
 
-    FlightClass response =
+    LoyaltyClass response =
         service
-            .flightclass()
+            .loyaltyclass()
             .patch(String.format("%s.%s", issuerId, classSuffix), patchBody)
             .execute();
 
@@ -247,17 +249,16 @@ public class DemoFlight {
    * @param header The message header.
    * @param body The message body.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String AddClassMessage(String issuerId, String classSuffix, String header, String body)
+  public String addClassMessage(String issuerId, String classSuffix, String header, String body)
       throws IOException {
     // Check if the class exists
     try {
-      service.flightclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.loyaltyclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -269,9 +270,9 @@ public class DemoFlight {
     AddMessageRequest message =
         new AddMessageRequest().setMessage(new Message().setHeader(header).setBody(body));
 
-    FlightClassAddMessageResponse response =
+    LoyaltyClassAddMessageResponse response =
         service
-            .flightclass()
+            .loyaltyclass()
             .addmessage(String.format("%s.%s", issuerId, classSuffix), message)
             .execute();
 
@@ -290,15 +291,14 @@ public class DemoFlight {
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String CreateObject(String issuerId, String classSuffix, String objectSuffix)
+  public String createObject(String issuerId, String classSuffix, String objectSuffix)
       throws IOException {
     // Check if the object exists
     try {
-      service.flightobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
 
-      System.out.println(String.format("Object %s.%s already exists!", issuerId, objectSuffix));
+      System.out.printf("Object %s.%s already exists!%n", issuerId, objectSuffix);
       return String.format("%s.%s", issuerId, objectSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() != 404) {
@@ -309,9 +309,9 @@ public class DemoFlight {
     }
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/tickets/boarding-passes/rest/v1/flightobject
-    FlightObject newObject =
-        new FlightObject()
+    // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyobject
+    LoyaltyObject newObject =
+        new LoyaltyObject()
             .setId(String.format("%s.%s", issuerId, objectSuffix))
             .setClassId(String.format("%s.%s", issuerId, classSuffix))
             .setState("ACTIVE")
@@ -328,11 +328,11 @@ public class DemoFlight {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -346,33 +346,35 @@ public class DemoFlight {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
-            .setPassengerName("Passenger name")
-            .setBoardingAndSeatingInfo(
-                new BoardingAndSeatingInfo().setBoardingGroup("B").setSeatNumber("42"))
-            .setReservationInfo(new ReservationInfo().setConfirmationCode("Confirmation code"));
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
+            .setAccountId("Account ID")
+            .setAccountName("Account name")
+            .setLoyaltyPoints(
+                new LoyaltyPoints()
+                    .setLabel("Points")
+                    .setBalance(new LoyaltyPointsBalance().setInt(800)));
 
-    FlightObject response = service.flightobject().insert(newObject).execute();
+    LoyaltyObject response = service.loyaltyobject().insert(newObject).execute();
 
     System.out.println("Object insert response");
     System.out.println(response.toPrettyString());
@@ -390,19 +392,18 @@ public class DemoFlight {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String UpdateObject(String issuerId, String objectSuffix) throws IOException {
-    FlightObject updatedObject;
+  public String updateObject(String issuerId, String objectSuffix) throws IOException {
+    LoyaltyObject updatedObject;
 
     // Check if the object exists
     try {
       updatedObject =
-          service.flightobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+          service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -420,14 +421,14 @@ public class DemoFlight {
 
     if (updatedObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
-      updatedObject.setLinksModuleData(new LinksModuleData().setUris(Arrays.asList(newLink)));
+      updatedObject.setLinksModuleData(new LinksModuleData().setUris(List.of(newLink)));
     } else {
       updatedObject.getLinksModuleData().getUris().add(newLink);
     }
 
-    FlightObject response =
+    LoyaltyObject response =
         service
-            .flightobject()
+            .loyaltyobject()
             .update(String.format("%s.%s", issuerId, objectSuffix), updatedObject)
             .execute();
 
@@ -445,19 +446,18 @@ public class DemoFlight {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String PatchObject(String issuerId, String objectSuffix) throws IOException {
-    FlightObject existingObject;
+  public String patchObject(String issuerId, String objectSuffix) throws IOException {
+    LoyaltyObject existingObject;
 
     // Check if the object exists
     try {
       existingObject =
-          service.flightobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+          service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -473,7 +473,7 @@ public class DemoFlight {
             .setUri("https://developers.google.com/wallet")
             .setDescription("New link description");
 
-    FlightObject patchBody = new FlightObject();
+    LoyaltyObject patchBody = new LoyaltyObject();
 
     if (existingObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
@@ -483,9 +483,9 @@ public class DemoFlight {
     }
     patchBody.getLinksModuleData().getUris().add(newLink);
 
-    FlightObject response =
+    LoyaltyObject response =
         service
-            .flightobject()
+            .loyaltyobject()
             .patch(String.format("%s.%s", issuerId, objectSuffix), patchBody)
             .execute();
 
@@ -506,16 +506,15 @@ public class DemoFlight {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String ExpireObject(String issuerId, String objectSuffix) throws IOException {
+  public String expireObject(String issuerId, String objectSuffix) throws IOException {
     // Check if the object exists
     try {
-      service.flightobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -525,11 +524,11 @@ public class DemoFlight {
     }
 
     // Patch the object, setting the pass as expired
-    FlightObject patchBody = new FlightObject().setState("EXPIRED");
+    LoyaltyObject patchBody = new LoyaltyObject().setState("EXPIRED");
 
-    FlightObject response =
+    LoyaltyObject response =
         service
-            .flightobject()
+            .loyaltyobject()
             .patch(String.format("%s.%s", issuerId, objectSuffix), patchBody)
             .execute();
 
@@ -549,17 +548,16 @@ public class DemoFlight {
    * @param header The message header.
    * @param body The message body.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String AddObjectMessage(String issuerId, String objectSuffix, String header, String body)
+  public String addObjectMessage(String issuerId, String objectSuffix, String header, String body)
       throws IOException {
     // Check if the object exists
     try {
-      service.flightobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -571,9 +569,9 @@ public class DemoFlight {
     AddMessageRequest message =
         new AddMessageRequest().setMessage(new Message().setHeader(header).setBody(body));
 
-    FlightObjectAddMessageResponse response =
+    LoyaltyObjectAddMessageResponse response =
         service
-            .flightobject()
+            .loyaltyobject()
             .addmessage(String.format("%s.%s", issuerId, objectSuffix), message)
             .execute();
 
@@ -597,27 +595,32 @@ public class DemoFlight {
    * @param objectSuffix Developer-defined unique ID for the pass object.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
+  public String createJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/tickets/boarding-passes/rest/v1/flightclass
-    FlightClass newClass =
-        new FlightClass()
+    // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyclass
+    LoyaltyClass newClass =
+        new LoyaltyClass()
             .setId(String.format("%s.%s", issuerId, classSuffix))
             .setIssuerName("Issuer name")
             .setReviewStatus("UNDER_REVIEW")
-            .setLocalScheduledDepartureDateTime("2023-07-02T15:30:00")
-            .setFlightHeader(
-                new FlightHeader()
-                    .setCarrier(new FlightCarrier().setCarrierIataCode("LX"))
-                    .setFlightNumber("123"))
-            .setOrigin(new AirportInfo().setAirportIataCode("LAX").setTerminal("1").setGate("A2"))
-            .setDestination(
-                new AirportInfo().setAirportIataCode("SFO").setTerminal("2").setGate("C3"));
+            .setProgramName("Program name")
+            .setProgramLogo(
+                new Image()
+                    .setSourceUri(
+                        new ImageUri()
+                            .setUri(
+                                "https://farm4.staticflickr.com/3723/11177041115_6e6a3b6f49_o.jpg"))
+                    .setContentDescription(
+                        new LocalizedString()
+                            .setDefaultValue(
+                                new TranslatedString()
+                                    .setLanguage("en-US")
+                                    .setValue("Logo description"))));
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/tickets/boarding-passes/rest/v1/flightobject
-    FlightObject newObject =
-        new FlightObject()
+    // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyobject
+    LoyaltyObject newObject =
+        new LoyaltyObject()
             .setId(String.format("%s.%s", issuerId, objectSuffix))
             .setClassId(String.format("%s.%s", issuerId, classSuffix))
             .setState("ACTIVE")
@@ -634,11 +637,11 @@ public class DemoFlight {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -652,43 +655,45 @@ public class DemoFlight {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
-            .setPassengerName("Passenger name")
-            .setBoardingAndSeatingInfo(
-                new BoardingAndSeatingInfo().setBoardingGroup("B").setSeatNumber("42"))
-            .setReservationInfo(new ReservationInfo().setConfirmationCode("Confirmation code"));
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
+            .setAccountId("Account ID")
+            .setAccountName("Account name")
+            .setLoyaltyPoints(
+                new LoyaltyPoints()
+                    .setLabel("Points")
+                    .setBalance(new LoyaltyPointsBalance().setInt(800)));
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
 
     // Create the Google Wallet payload and add to the JWT
     HashMap<String, Object> payload = new HashMap<String, Object>();
-    payload.put("flightClasses", Arrays.asList(newClass));
-    payload.put("flightObjects", Arrays.asList(newObject));
+    payload.put("loyaltyClasses", List.of(newClass));
+    payload.put("loyaltyObjects", List.of(newObject));
     claims.put("payload", payload);
 
     // The service account credentials are used to sign the JWT
@@ -698,7 +703,7 @@ public class DemoFlight {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -719,7 +724,7 @@ public class DemoFlight {
    * @param issuerId The issuer ID being used for this request.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTExistingObjects(String issuerId) {
+  public String createJWTExistingObjects(String issuerId) {
     // Multiple pass types can be added at the same time
     // At least one type must be specified in the JWT claims
     // Note: Make sure to replace the placeholder class and object suffixes
@@ -728,64 +733,64 @@ public class DemoFlight {
     // Event tickets
     objectsToAdd.put(
         "eventTicketObjects",
-        Arrays.asList(
-            new EventTicketObject()
-                .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
+            List.of(
+                    new EventTicketObject()
+                            .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
 
     // Boarding passes
     objectsToAdd.put(
         "flightObjects",
-        Arrays.asList(
-            new FlightObject()
-                .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
+            List.of(
+                    new FlightObject()
+                            .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
 
     // Generic passes
     objectsToAdd.put(
         "genericObjects",
-        Arrays.asList(
-            new GenericObject()
-                .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
+            List.of(
+                    new GenericObject()
+                            .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
 
     // Gift cards
     objectsToAdd.put(
         "giftCardObjects",
-        Arrays.asList(
-            new GiftCardObject()
-                .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
+            List.of(
+                    new GiftCardObject()
+                            .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
 
     // Loyalty cards
     objectsToAdd.put(
         "loyaltyObjects",
-        Arrays.asList(
-            new LoyaltyObject()
-                .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
+            List.of(
+                    new LoyaltyObject()
+                            .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
 
     // Offers
     objectsToAdd.put(
         "offerObjects",
-        Arrays.asList(
-            new OfferObject()
-                .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
+            List.of(
+                    new OfferObject()
+                            .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
 
     // Transit passes
     objectsToAdd.put(
         "transitObjects",
-        Arrays.asList(
-            new TransitObject()
-                .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
+            List.of(
+                    new TransitObject()
+                            .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
     claims.put("payload", objectsToAdd);
 
@@ -796,7 +801,7 @@ public class DemoFlight {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -808,17 +813,16 @@ public class DemoFlight {
    *
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
-   * @throws IOException
    */
   public void BatchCreateObjects(String issuerId, String classSuffix) throws IOException {
     // Create the batch request client
     BatchRequest batch = service.batch(new HttpCredentialsAdapter(credentials));
 
     // The callback will be invoked for each request in the batch
-    JsonBatchCallback<FlightObject> callback =
-        new JsonBatchCallback<FlightObject>() {
+    JsonBatchCallback<LoyaltyObject> callback =
+        new JsonBatchCallback<LoyaltyObject>() {
           // Invoked if the request was successful
-          public void onSuccess(FlightObject response, HttpHeaders responseHeaders) {
+          public void onSuccess(LoyaltyObject response, HttpHeaders responseHeaders) {
             System.out.println("Batch insert response");
             System.out.println(response.toString());
           }
@@ -835,9 +839,9 @@ public class DemoFlight {
       String objectSuffix = UUID.randomUUID().toString().replaceAll("[^\\w.-]", "_");
 
       // See link below for more information on required properties
-      // https://developers.google.com/wallet/tickets/boarding-passes/rest/v1/flightobject
-      FlightObject batchObject =
-          new FlightObject()
+      // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyobject
+      LoyaltyObject batchObject =
+          new LoyaltyObject()
               .setId(String.format("%s.%s", issuerId, objectSuffix))
               .setClassId(String.format("%s.%s", issuerId, classSuffix))
               .setState("ACTIVE")
@@ -854,11 +858,11 @@ public class DemoFlight {
                                       .setLanguage("en-US")
                                       .setValue("Hero image description"))))
               .setTextModulesData(
-                  Arrays.asList(
-                      new TextModuleData()
-                          .setHeader("Text module header")
-                          .setBody("Text module body")
-                          .setId("TEXT_MODULE_ID")))
+                      List.of(
+                              new TextModuleData()
+                                      .setHeader("Text module header")
+                                      .setBody("Text module body")
+                                      .setId("TEXT_MODULE_ID")))
               .setLinksModuleData(
                   new LinksModuleData()
                       .setUris(
@@ -872,33 +876,35 @@ public class DemoFlight {
                                   .setDescription("Link module tel description")
                                   .setId("LINK_MODULE_TEL_ID"))))
               .setImageModulesData(
-                  Arrays.asList(
-                      new ImageModuleData()
-                          .setMainImage(
-                              new Image()
-                                  .setSourceUri(
-                                      new ImageUri()
-                                          .setUri(
-                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                  .setContentDescription(
-                                      new LocalizedString()
-                                          .setDefaultValue(
-                                              new TranslatedString()
-                                                  .setLanguage("en-US")
-                                                  .setValue("Image module description"))))
-                          .setId("IMAGE_MODULE_ID")))
+                      List.of(
+                              new ImageModuleData()
+                                      .setMainImage(
+                                              new Image()
+                                                      .setSourceUri(
+                                                              new ImageUri()
+                                                                      .setUri(
+                                                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                      .setContentDescription(
+                                                              new LocalizedString()
+                                                                      .setDefaultValue(
+                                                                              new TranslatedString()
+                                                                                      .setLanguage("en-US")
+                                                                                      .setValue("Image module description"))))
+                                      .setId("IMAGE_MODULE_ID")))
               .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
               .setLocations(
-                  Arrays.asList(
-                      new LatLongPoint()
-                          .setLatitude(37.424015499999996)
-                          .setLongitude(-122.09259560000001)))
-              .setPassengerName("Passenger name")
-              .setBoardingAndSeatingInfo(
-                  new BoardingAndSeatingInfo().setBoardingGroup("B").setSeatNumber("42"))
-              .setReservationInfo(new ReservationInfo().setConfirmationCode("Confirmation code"));
+                      List.of(
+                              new LatLongPoint()
+                                      .setLatitude(37.424015499999996)
+                                      .setLongitude(-122.09259560000001)))
+              .setAccountId("Account ID")
+              .setAccountName("Account name")
+              .setLoyaltyPoints(
+                  new LoyaltyPoints()
+                      .setLabel("Points")
+                      .setBalance(new LoyaltyPointsBalance().setInt(800)));
 
-      service.flightobject().insert(batchObject).queue(batch, callback);
+      service.loyaltyobject().insert(batchObject).queue(batch, callback);
     }
 
     // Invoke the batch API calls
