+++ java/src/main/java/com/google/developers/wallet/rest/DemoOffer.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.google.developers.wallet.rest;
 
 // [START setup]
 // [START imports]
@@ -35,8 +36,8 @@ import java.security.interfaces.RSAPrivateKey;
 import java.util.*;
 // [END imports]
 
-/** Demo class for creating and managing Loyalty cards in Google Wallet. */
-public class DemoLoyalty {
+/** Demo class for creating and managing Offers in Google Wallet. */
+public class DemoOffer {
   /**
    * Path to service account key file from Google Cloud Console. Environment variable:
    * GOOGLE_APPLICATION_CREDENTIALS.
@@ -49,11 +50,11 @@ public class DemoLoyalty {
   /** Google Wallet service client. */
   public static Walletobjects service;
 
-  public DemoLoyalty() throws Exception {
+  public DemoOffer() throws Exception {
     keyFilePath =
         System.getenv().getOrDefault("GOOGLE_APPLICATION_CREDENTIALS", "/path/to/key.json");
 
-    Auth();
+    auth();
   }
   // [END setup]
 
@@ -61,14 +62,13 @@ public class DemoLoyalty {
   /**
    * Create authenticated HTTP client using a service account file.
    *
-   * @throws Exception
    */
-  public void Auth() throws Exception {
+  public void auth() throws Exception {
     String scope = "https://www.googleapis.com/auth/wallet_object.issuer";
 
     credentials =
         GoogleCredentials.fromStream(new FileInputStream(keyFilePath))
-            .createScoped(Arrays.asList(scope));
+            .createScoped(List.of(scope));
     credentials.refresh();
 
     HttpTransport httpTransport = GoogleNetHttpTransport.newTrustedTransport();
@@ -91,14 +91,13 @@ public class DemoLoyalty {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String CreateClass(String issuerId, String classSuffix) throws IOException {
+  public String createClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
-      service.loyaltyclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.offerclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
 
-      System.out.println(String.format("Class %s.%s already exists!", issuerId, classSuffix));
+      System.out.printf("Class %s.%s already exists!%n", issuerId, classSuffix);
       return String.format("%s.%s", issuerId, classSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() != 404) {
@@ -109,27 +108,17 @@ public class DemoLoyalty {
     }
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyclass
-    LoyaltyClass newClass =
-        new LoyaltyClass()
+    // https://developers.google.com/wallet/retail/offers/rest/v1/offerclass
+    OfferClass newClass =
+        new OfferClass()
             .setId(String.format("%s.%s", issuerId, classSuffix))
             .setIssuerName("Issuer name")
             .setReviewStatus("UNDER_REVIEW")
-            .setProgramName("Program name")
-            .setProgramLogo(
-                new Image()
-                    .setSourceUri(
-                        new ImageUri()
-                            .setUri(
-                                "https://farm4.staticflickr.com/3723/11177041115_6e6a3b6f49_o.jpg"))
-                    .setContentDescription(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString()
-                                    .setLanguage("en-US")
-                                    .setValue("Logo description"))));
+            .setProvider("Provider name")
+            .setTitle("Offer title")
+            .setRedemptionChannel("ONLINE");
 
-    LoyaltyClass response = service.loyaltyclass().insert(newClass).execute();
+    OfferClass response = service.offerclass().insert(newClass).execute();
 
     System.out.println("Class insert response");
     System.out.println(response.toPrettyString());
@@ -147,19 +136,18 @@ public class DemoLoyalty {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String UpdateClass(String issuerId, String classSuffix) throws IOException {
-    LoyaltyClass updatedClass;
+  public String updateClass(String issuerId, String classSuffix) throws IOException {
+    OfferClass updatedClass;
 
     // Check if the class exists
     try {
       updatedClass =
-          service.loyaltyclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+          service.offerclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -178,9 +166,9 @@ public class DemoLoyalty {
     // Note: reviewStatus must be 'UNDER_REVIEW' or 'DRAFT' for updates
     updatedClass.setReviewStatus("UNDER_REVIEW");
 
-    LoyaltyClass response =
+    OfferClass response =
         service
-            .loyaltyclass()
+            .offerclass()
             .update(String.format("%s.%s", issuerId, classSuffix), updatedClass)
             .execute();
 
@@ -200,16 +188,15 @@ public class DemoLoyalty {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String PatchClass(String issuerId, String classSuffix) throws IOException {
+  public String patchClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
-      service.loyaltyclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.offerclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -220,8 +207,8 @@ public class DemoLoyalty {
 
     // Class exists
     // Patch the class by adding a homepage
-    LoyaltyClass patchBody =
-        new LoyaltyClass()
+    OfferClass patchBody =
+        new OfferClass()
             .setHomepageUri(
                 new Uri()
                     .setUri("https://developers.google.com/wallet")
@@ -230,9 +217,9 @@ public class DemoLoyalty {
             // Note: reviewStatus must be 'UNDER_REVIEW' or 'DRAFT' for updates
             .setReviewStatus("UNDER_REVIEW");
 
-    LoyaltyClass response =
+    OfferClass response =
         service
-            .loyaltyclass()
+            .offerclass()
             .patch(String.format("%s.%s", issuerId, classSuffix), patchBody)
             .execute();
 
@@ -252,17 +239,16 @@ public class DemoLoyalty {
    * @param header The message header.
    * @param body The message body.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String AddClassMessage(String issuerId, String classSuffix, String header, String body)
+  public String addClassMessage(String issuerId, String classSuffix, String header, String body)
       throws IOException {
     // Check if the class exists
     try {
-      service.loyaltyclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
+      service.offerclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -274,9 +260,9 @@ public class DemoLoyalty {
     AddMessageRequest message =
         new AddMessageRequest().setMessage(new Message().setHeader(header).setBody(body));
 
-    LoyaltyClassAddMessageResponse response =
+    OfferClassAddMessageResponse response =
         service
-            .loyaltyclass()
+            .offerclass()
             .addmessage(String.format("%s.%s", issuerId, classSuffix), message)
             .execute();
 
@@ -295,18 +281,20 @@ public class DemoLoyalty {
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String CreateObject(String issuerId, String classSuffix, String objectSuffix)
+  public String createObject(String issuerId, String classSuffix, String objectSuffix)
       throws IOException {
     // Check if the object exists
     try {
-      service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.offerobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
 
-      System.out.println(String.format("Object %s.%s already exists!", issuerId, objectSuffix));
+      System.out.printf("Object %s.%s already exists!%n", issuerId, objectSuffix);
       return String.format("%s.%s", issuerId, objectSuffix);
     } catch (GoogleJsonResponseException ex) {
-      if (ex.getStatusCode() != 404) {
+      if (ex.getStatusCode() == 404) {
+        // Object does not exist
+        // Do nothing
+      } else {
         // Something else went wrong...
         ex.printStackTrace();
         return String.format("%s.%s", issuerId, objectSuffix);
@@ -314,9 +302,9 @@ public class DemoLoyalty {
     }
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyobject
-    LoyaltyObject newObject =
-        new LoyaltyObject()
+    // https://developers.google.com/wallet/retail/offers/rest/v1/offerobject
+    OfferObject newObject =
+        new OfferObject()
             .setId(String.format("%s.%s", issuerId, objectSuffix))
             .setClassId(String.format("%s.%s", issuerId, classSuffix))
             .setState("ACTIVE")
@@ -333,11 +321,11 @@ public class DemoLoyalty {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -351,35 +339,33 @@ public class DemoLoyalty {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
-            .setAccountId("Account ID")
-            .setAccountName("Account name")
-            .setLoyaltyPoints(
-                new LoyaltyPoints()
-                    .setLabel("Points")
-                    .setBalance(new LoyaltyPointsBalance().setInt(800)));
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
+            .setValidTimeInterval(
+                new TimeInterval()
+                    .setStart(new DateTime().setDate("2023-06-12T23:20:50.52Z"))
+                    .setEnd(new DateTime().setDate("2023-12-12T23:20:50.52Z")));
 
-    LoyaltyObject response = service.loyaltyobject().insert(newObject).execute();
+    OfferObject response = service.offerobject().insert(newObject).execute();
 
     System.out.println("Object insert response");
     System.out.println(response.toPrettyString());
@@ -397,19 +383,18 @@ public class DemoLoyalty {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String UpdateObject(String issuerId, String objectSuffix) throws IOException {
-    LoyaltyObject updatedObject;
+  public String updateObject(String issuerId, String objectSuffix) throws IOException {
+    OfferObject updatedObject;
 
     // Check if the object exists
     try {
       updatedObject =
-          service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+          service.offerobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -427,14 +412,14 @@ public class DemoLoyalty {
 
     if (updatedObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
-      updatedObject.setLinksModuleData(new LinksModuleData().setUris(Arrays.asList(newLink)));
+      updatedObject.setLinksModuleData(new LinksModuleData().setUris(List.of(newLink)));
     } else {
       updatedObject.getLinksModuleData().getUris().add(newLink);
     }
 
-    LoyaltyObject response =
+    OfferObject response =
         service
-            .loyaltyobject()
+            .offerobject()
             .update(String.format("%s.%s", issuerId, objectSuffix), updatedObject)
             .execute();
 
@@ -452,19 +437,18 @@ public class DemoLoyalty {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String PatchObject(String issuerId, String objectSuffix) throws IOException {
-    LoyaltyObject existingObject;
+  public String patchObject(String issuerId, String objectSuffix) throws IOException {
+    OfferObject existingObject;
 
     // Check if the object exists
     try {
       existingObject =
-          service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+          service.offerobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -480,7 +464,7 @@ public class DemoLoyalty {
             .setUri("https://developers.google.com/wallet")
             .setDescription("New link description");
 
-    LoyaltyObject patchBody = new LoyaltyObject();
+    OfferObject patchBody = new OfferObject();
 
     if (existingObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
@@ -490,9 +474,9 @@ public class DemoLoyalty {
     }
     patchBody.getLinksModuleData().getUris().add(newLink);
 
-    LoyaltyObject response =
+    OfferObject response =
         service
-            .loyaltyobject()
+            .offerobject()
             .patch(String.format("%s.%s", issuerId, objectSuffix), patchBody)
             .execute();
 
@@ -513,16 +497,15 @@ public class DemoLoyalty {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String ExpireObject(String issuerId, String objectSuffix) throws IOException {
+  public String expireObject(String issuerId, String objectSuffix) throws IOException {
     // Check if the object exists
     try {
-      service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.offerobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -532,11 +515,11 @@ public class DemoLoyalty {
     }
 
     // Patch the object, setting the pass as expired
-    LoyaltyObject patchBody = new LoyaltyObject().setState("EXPIRED");
+    OfferObject patchBody = new OfferObject().setState("EXPIRED");
 
-    LoyaltyObject response =
+    OfferObject response =
         service
-            .loyaltyobject()
+            .offerobject()
             .patch(String.format("%s.%s", issuerId, objectSuffix), patchBody)
             .execute();
 
@@ -556,17 +539,16 @@ public class DemoLoyalty {
    * @param header The message header.
    * @param body The message body.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String AddObjectMessage(String issuerId, String objectSuffix, String header, String body)
+  public String addObjectMessage(String issuerId, String objectSuffix, String header, String body)
       throws IOException {
     // Check if the object exists
     try {
-      service.loyaltyobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
+      service.offerobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -578,9 +560,9 @@ public class DemoLoyalty {
     AddMessageRequest message =
         new AddMessageRequest().setMessage(new Message().setHeader(header).setBody(body));
 
-    LoyaltyObjectAddMessageResponse response =
+    OfferObjectAddMessageResponse response =
         service
-            .loyaltyobject()
+            .offerobject()
             .addmessage(String.format("%s.%s", issuerId, objectSuffix), message)
             .execute();
 
@@ -604,32 +586,22 @@ public class DemoLoyalty {
    * @param objectSuffix Developer-defined unique ID for the pass object.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
+  public String createJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyclass
-    LoyaltyClass newClass =
-        new LoyaltyClass()
+    // https://developers.google.com/wallet/retail/offers/rest/v1/offerclass
+    OfferClass newClass =
+        new OfferClass()
             .setId(String.format("%s.%s", issuerId, classSuffix))
             .setIssuerName("Issuer name")
             .setReviewStatus("UNDER_REVIEW")
-            .setProgramName("Program name")
-            .setProgramLogo(
-                new Image()
-                    .setSourceUri(
-                        new ImageUri()
-                            .setUri(
-                                "https://farm4.staticflickr.com/3723/11177041115_6e6a3b6f49_o.jpg"))
-                    .setContentDescription(
-                        new LocalizedString()
-                            .setDefaultValue(
-                                new TranslatedString()
-                                    .setLanguage("en-US")
-                                    .setValue("Logo description"))));
+            .setProvider("Provider name")
+            .setTitle("Offer title")
+            .setRedemptionChannel("ONLINE");
 
     // See link below for more information on required properties
-    // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyobject
-    LoyaltyObject newObject =
-        new LoyaltyObject()
+    // https://developers.google.com/wallet/retail/offers/rest/v1/offerobject
+    OfferObject newObject =
+        new OfferObject()
             .setId(String.format("%s.%s", issuerId, objectSuffix))
             .setClassId(String.format("%s.%s", issuerId, classSuffix))
             .setState("ACTIVE")
@@ -646,11 +618,11 @@ public class DemoLoyalty {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -664,45 +636,43 @@ public class DemoLoyalty {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
-            .setAccountId("Account ID")
-            .setAccountName("Account name")
-            .setLoyaltyPoints(
-                new LoyaltyPoints()
-                    .setLabel("Points")
-                    .setBalance(new LoyaltyPointsBalance().setInt(800)));
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
+            .setValidTimeInterval(
+                new TimeInterval()
+                    .setStart(new DateTime().setDate("2023-06-12T23:20:50.52Z"))
+                    .setEnd(new DateTime().setDate("2023-12-12T23:20:50.52Z")));
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
 
     // Create the Google Wallet payload and add to the JWT
     HashMap<String, Object> payload = new HashMap<String, Object>();
-    payload.put("loyaltyClasses", Arrays.asList(newClass));
-    payload.put("loyaltyObjects", Arrays.asList(newObject));
+    payload.put("offerClasses", List.of(newClass));
+    payload.put("offerObjects", List.of(newObject));
     claims.put("payload", payload);
 
     // The service account credentials are used to sign the JWT
@@ -712,7 +682,7 @@ public class DemoLoyalty {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -733,7 +703,7 @@ public class DemoLoyalty {
    * @param issuerId The issuer ID being used for this request.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTExistingObjects(String issuerId) {
+  public String createJWTExistingObjects(String issuerId) {
     // Multiple pass types can be added at the same time
     // At least one type must be specified in the JWT claims
     // Note: Make sure to replace the placeholder class and object suffixes
@@ -742,64 +712,64 @@ public class DemoLoyalty {
     // Event tickets
     objectsToAdd.put(
         "eventTicketObjects",
-        Arrays.asList(
-            new EventTicketObject()
-                .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
+            List.of(
+                    new EventTicketObject()
+                            .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
 
     // Boarding passes
     objectsToAdd.put(
         "flightObjects",
-        Arrays.asList(
-            new FlightObject()
-                .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
+            List.of(
+                    new FlightObject()
+                            .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
 
     // Generic passes
     objectsToAdd.put(
         "genericObjects",
-        Arrays.asList(
-            new GenericObject()
-                .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
+            List.of(
+                    new GenericObject()
+                            .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
 
     // Gift cards
     objectsToAdd.put(
         "giftCardObjects",
-        Arrays.asList(
-            new GiftCardObject()
-                .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
+            List.of(
+                    new GiftCardObject()
+                            .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
 
     // Loyalty cards
     objectsToAdd.put(
         "loyaltyObjects",
-        Arrays.asList(
-            new LoyaltyObject()
-                .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
+            List.of(
+                    new LoyaltyObject()
+                            .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
 
     // Offers
     objectsToAdd.put(
         "offerObjects",
-        Arrays.asList(
-            new OfferObject()
-                .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
+            List.of(
+                    new OfferObject()
+                            .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
 
     // Transit passes
     objectsToAdd.put(
         "transitObjects",
-        Arrays.asList(
-            new TransitObject()
-                .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
+            List.of(
+                    new TransitObject()
+                            .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
     claims.put("payload", objectsToAdd);
 
@@ -810,7 +780,7 @@ public class DemoLoyalty {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -822,17 +792,16 @@ public class DemoLoyalty {
    *
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
-   * @throws IOException
    */
-  public void BatchCreateObjects(String issuerId, String classSuffix) throws IOException {
+  public void batchCreateObjects(String issuerId, String classSuffix) throws IOException {
     // Create the batch request client
     BatchRequest batch = service.batch(new HttpCredentialsAdapter(credentials));
 
     // The callback will be invoked for each request in the batch
-    JsonBatchCallback<LoyaltyObject> callback =
-        new JsonBatchCallback<LoyaltyObject>() {
+    JsonBatchCallback<OfferObject> callback =
+        new JsonBatchCallback<OfferObject>() {
           // Invoked if the request was successful
-          public void onSuccess(LoyaltyObject response, HttpHeaders responseHeaders) {
+          public void onSuccess(OfferObject response, HttpHeaders responseHeaders) {
             System.out.println("Batch insert response");
             System.out.println(response.toString());
           }
@@ -849,9 +818,9 @@ public class DemoLoyalty {
       String objectSuffix = UUID.randomUUID().toString().replaceAll("[^\\w.-]", "_");
 
       // See link below for more information on required properties
-      // https://developers.google.com/wallet/retail/loyalty-cards/rest/v1/loyaltyobject
-      LoyaltyObject batchObject =
-          new LoyaltyObject()
+      // https://developers.google.com/wallet/retail/offers/rest/v1/offerobject
+      OfferObject batchObject =
+          new OfferObject()
               .setId(String.format("%s.%s", issuerId, objectSuffix))
               .setClassId(String.format("%s.%s", issuerId, classSuffix))
               .setState("ACTIVE")
@@ -868,11 +837,11 @@ public class DemoLoyalty {
                                       .setLanguage("en-US")
                                       .setValue("Hero image description"))))
               .setTextModulesData(
-                  Arrays.asList(
-                      new TextModuleData()
-                          .setHeader("Text module header")
-                          .setBody("Text module body")
-                          .setId("TEXT_MODULE_ID")))
+                      List.of(
+                              new TextModuleData()
+                                      .setHeader("Text module header")
+                                      .setBody("Text module body")
+                                      .setId("TEXT_MODULE_ID")))
               .setLinksModuleData(
                   new LinksModuleData()
                       .setUris(
@@ -886,35 +855,33 @@ public class DemoLoyalty {
                                   .setDescription("Link module tel description")
                                   .setId("LINK_MODULE_TEL_ID"))))
               .setImageModulesData(
-                  Arrays.asList(
-                      new ImageModuleData()
-                          .setMainImage(
-                              new Image()
-                                  .setSourceUri(
-                                      new ImageUri()
-                                          .setUri(
-                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                  .setContentDescription(
-                                      new LocalizedString()
-                                          .setDefaultValue(
-                                              new TranslatedString()
-                                                  .setLanguage("en-US")
-                                                  .setValue("Image module description"))))
-                          .setId("IMAGE_MODULE_ID")))
+                      List.of(
+                              new ImageModuleData()
+                                      .setMainImage(
+                                              new Image()
+                                                      .setSourceUri(
+                                                              new ImageUri()
+                                                                      .setUri(
+                                                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                      .setContentDescription(
+                                                              new LocalizedString()
+                                                                      .setDefaultValue(
+                                                                              new TranslatedString()
+                                                                                      .setLanguage("en-US")
+                                                                                      .setValue("Image module description"))))
+                                      .setId("IMAGE_MODULE_ID")))
               .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
               .setLocations(
-                  Arrays.asList(
-                      new LatLongPoint()
-                          .setLatitude(37.424015499999996)
-                          .setLongitude(-122.09259560000001)))
-              .setAccountId("Account ID")
-              .setAccountName("Account name")
-              .setLoyaltyPoints(
-                  new LoyaltyPoints()
-                      .setLabel("Points")
-                      .setBalance(new LoyaltyPointsBalance().setInt(800)));
+                      List.of(
+                              new LatLongPoint()
+                                      .setLatitude(37.424015499999996)
+                                      .setLongitude(-122.09259560000001)))
+              .setValidTimeInterval(
+                  new TimeInterval()
+                      .setStart(new DateTime().setDate("2023-06-12T23:20:50.52Z"))
+                      .setEnd(new DateTime().setDate("2023-12-12T23:20:50.52Z")));
 
-      service.loyaltyobject().insert(batchObject).queue(batch, callback);
+      service.offerobject().insert(batchObject).queue(batch, callback);
     }
 
     // Invoke the batch API calls
