+++ java/src/main/java/com/google/developers/wallet/rest/DemoTransit.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.google.developers.wallet.rest;
 
 // [START setup]
 // [START imports]
@@ -52,7 +53,7 @@ public class DemoTransit {
     keyFilePath =
         System.getenv().getOrDefault("GOOGLE_APPLICATION_CREDENTIALS", "/path/to/key.json");
 
-    Auth();
+    auth();
   }
   // [END setup]
 
@@ -60,14 +61,13 @@ public class DemoTransit {
   /**
    * Create authenticated HTTP client using a service account file.
    *
-   * @throws Exception
    */
-  public void Auth() throws Exception {
+  public void auth() throws Exception {
     String scope = "https://www.googleapis.com/auth/wallet_object.issuer";
 
     credentials =
         GoogleCredentials.fromStream(new FileInputStream(keyFilePath))
-            .createScoped(Arrays.asList(scope));
+            .createScoped(List.of(scope));
     credentials.refresh();
 
     HttpTransport httpTransport = GoogleNetHttpTransport.newTrustedTransport();
@@ -90,14 +90,13 @@ public class DemoTransit {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String CreateClass(String issuerId, String classSuffix) throws IOException {
+  public String createClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
       service.transitclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
 
-      System.out.println(String.format("Class %s.%s already exists!", issuerId, classSuffix));
+      System.out.printf("Class %s.%s already exists!%n", issuerId, classSuffix);
       return String.format("%s.%s", issuerId, classSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() != 404) {
@@ -146,9 +145,8 @@ public class DemoTransit {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String UpdateClass(String issuerId, String classSuffix) throws IOException {
+  public String updateClass(String issuerId, String classSuffix) throws IOException {
     TransitClass updatedClass;
 
     // Check if the class exists
@@ -158,7 +156,7 @@ public class DemoTransit {
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -199,16 +197,15 @@ public class DemoTransit {
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String PatchClass(String issuerId, String classSuffix) throws IOException {
+  public String patchClass(String issuerId, String classSuffix) throws IOException {
     // Check if the class exists
     try {
       service.transitclass().get(String.format("%s.%s", issuerId, classSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -251,9 +248,8 @@ public class DemoTransit {
    * @param header The message header.
    * @param body The message body.
    * @return The pass class ID: "{issuerId}.{classSuffix}"
-   * @throws IOException
    */
-  public String AddClassMessage(String issuerId, String classSuffix, String header, String body)
+  public String addClassMessage(String issuerId, String classSuffix, String header, String body)
       throws IOException {
     // Check if the class exists
     try {
@@ -261,7 +257,7 @@ public class DemoTransit {
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Class does not exist
-        System.out.println(String.format("Class %s.%s not found!", issuerId, classSuffix));
+        System.out.printf("Class %s.%s not found!%n", issuerId, classSuffix);
         return String.format("%s.%s", issuerId, classSuffix);
       } else {
         // Something else went wrong...
@@ -294,15 +290,14 @@ public class DemoTransit {
    * @param classSuffix Developer-defined unique ID for this pass class.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String CreateObject(String issuerId, String classSuffix, String objectSuffix)
+  public String createObject(String issuerId, String classSuffix, String objectSuffix)
       throws IOException {
     // Check if the object exists
     try {
       service.transitobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
 
-      System.out.println(String.format("Object %s.%s already exists!", issuerId, objectSuffix));
+      System.out.printf("Object %s.%s already exists!%n", issuerId, objectSuffix);
       return String.format("%s.%s", issuerId, objectSuffix);
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
@@ -335,11 +330,11 @@ public class DemoTransit {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -353,27 +348,27 @@ public class DemoTransit {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
             .setPassengerType("SINGLE_PASSENGER")
             .setPassengerNames("Passenger names")
             .setTripType("ONE_WAY")
@@ -420,9 +415,8 @@ public class DemoTransit {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String UpdateObject(String issuerId, String objectSuffix) throws IOException {
+  public String updateObject(String issuerId, String objectSuffix) throws IOException {
     TransitObject updatedObject;
 
     // Check if the object exists
@@ -432,7 +426,7 @@ public class DemoTransit {
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -450,7 +444,7 @@ public class DemoTransit {
 
     if (updatedObject.getLinksModuleData() == null) {
       // LinksModuleData was not set on the original object
-      updatedObject.setLinksModuleData(new LinksModuleData().setUris(Arrays.asList(newLink)));
+      updatedObject.setLinksModuleData(new LinksModuleData().setUris(List.of(newLink)));
     } else {
       updatedObject.getLinksModuleData().getUris().add(newLink);
     }
@@ -475,9 +469,8 @@ public class DemoTransit {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String PatchObject(String issuerId, String objectSuffix) throws IOException {
+  public String patchObject(String issuerId, String objectSuffix) throws IOException {
     TransitObject existingObject;
 
     // Check if the object exists
@@ -487,7 +480,7 @@ public class DemoTransit {
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -536,16 +529,15 @@ public class DemoTransit {
    * @param issuerId The issuer ID being used for this request.
    * @param objectSuffix Developer-defined unique ID for this pass object.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String ExpireObject(String issuerId, String objectSuffix) throws IOException {
+  public String expireObject(String issuerId, String objectSuffix) throws IOException {
     // Check if the object exists
     try {
       service.transitobject().get(String.format("%s.%s", issuerId, objectSuffix)).execute();
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -579,9 +571,8 @@ public class DemoTransit {
    * @param header The message header.
    * @param body The message body.
    * @return The pass object ID: "{issuerId}.{objectSuffix}"
-   * @throws IOException
    */
-  public String AddObjectMessage(String issuerId, String objectSuffix, String header, String body)
+  public String addObjectMessage(String issuerId, String objectSuffix, String header, String body)
       throws IOException {
     // Check if the object exists
     try {
@@ -589,7 +580,7 @@ public class DemoTransit {
     } catch (GoogleJsonResponseException ex) {
       if (ex.getStatusCode() == 404) {
         // Object does not exist
-        System.out.println(String.format("Object %s.%s not found!", issuerId, objectSuffix));
+        System.out.printf("Object %s.%s not found!%n", issuerId, objectSuffix);
         return String.format("%s.%s", issuerId, objectSuffix);
       } else {
         // Something else went wrong...
@@ -627,7 +618,7 @@ public class DemoTransit {
    * @param objectSuffix Developer-defined unique ID for the pass object.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
+  public String createJWTNewObjects(String issuerId, String classSuffix, String objectSuffix) {
     // See link below for more information on required properties
     // https://developers.google.com/wallet/tickets/transit-passes/qr-code/rest/v1/transitclass
     TransitClass newClass =
@@ -669,11 +660,11 @@ public class DemoTransit {
                                     .setLanguage("en-US")
                                     .setValue("Hero image description"))))
             .setTextModulesData(
-                Arrays.asList(
-                    new TextModuleData()
-                        .setHeader("Text module header")
-                        .setBody("Text module body")
-                        .setId("TEXT_MODULE_ID")))
+                    List.of(
+                            new TextModuleData()
+                                    .setHeader("Text module header")
+                                    .setBody("Text module body")
+                                    .setId("TEXT_MODULE_ID")))
             .setLinksModuleData(
                 new LinksModuleData()
                     .setUris(
@@ -687,27 +678,27 @@ public class DemoTransit {
                                 .setDescription("Link module tel description")
                                 .setId("LINK_MODULE_TEL_ID"))))
             .setImageModulesData(
-                Arrays.asList(
-                    new ImageModuleData()
-                        .setMainImage(
-                            new Image()
-                                .setSourceUri(
-                                    new ImageUri()
-                                        .setUri(
-                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                .setContentDescription(
-                                    new LocalizedString()
-                                        .setDefaultValue(
-                                            new TranslatedString()
-                                                .setLanguage("en-US")
-                                                .setValue("Image module description"))))
-                        .setId("IMAGE_MODULE_ID")))
+                    List.of(
+                            new ImageModuleData()
+                                    .setMainImage(
+                                            new Image()
+                                                    .setSourceUri(
+                                                            new ImageUri()
+                                                                    .setUri(
+                                                                            "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                    .setContentDescription(
+                                                            new LocalizedString()
+                                                                    .setDefaultValue(
+                                                                            new TranslatedString()
+                                                                                    .setLanguage("en-US")
+                                                                                    .setValue("Image module description"))))
+                                    .setId("IMAGE_MODULE_ID")))
             .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
             .setLocations(
-                Arrays.asList(
-                    new LatLongPoint()
-                        .setLatitude(37.424015499999996)
-                        .setLongitude(-122.09259560000001)))
+                    List.of(
+                            new LatLongPoint()
+                                    .setLatitude(37.424015499999996)
+                                    .setLongitude(-122.09259560000001)))
             .setPassengerType("SINGLE_PASSENGER")
             .setPassengerNames("Passenger names")
             .setTripType("ONE_WAY")
@@ -740,13 +731,13 @@ public class DemoTransit {
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
 
     // Create the Google Wallet payload and add to the JWT
     HashMap<String, Object> payload = new HashMap<String, Object>();
-    payload.put("transitClasses", Arrays.asList(newClass));
-    payload.put("transitObjects", Arrays.asList(newObject));
+    payload.put("transitClasses", List.of(newClass));
+    payload.put("transitObjects", List.of(newObject));
     claims.put("payload", payload);
 
     // The service account credentials are used to sign the JWT
@@ -756,7 +747,7 @@ public class DemoTransit {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -777,7 +768,7 @@ public class DemoTransit {
    * @param issuerId The issuer ID being used for this request.
    * @return An "Add to Google Wallet" link.
    */
-  public String CreateJWTExistingObjects(String issuerId) {
+  public String createJWTExistingObjects(String issuerId) {
     // Multiple pass types can be added at the same time
     // At least one type must be specified in the JWT claims
     // Note: Make sure to replace the placeholder class and object suffixes
@@ -786,64 +777,64 @@ public class DemoTransit {
     // Event tickets
     objectsToAdd.put(
         "eventTicketObjects",
-        Arrays.asList(
-            new EventTicketObject()
-                .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
+            List.of(
+                    new EventTicketObject()
+                            .setId(String.format("%s.%s", issuerId, "EVENT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "EVENT_CLASS_SUFFIX"))));
 
     // Boarding passes
     objectsToAdd.put(
         "flightObjects",
-        Arrays.asList(
-            new FlightObject()
-                .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
+            List.of(
+                    new FlightObject()
+                            .setId(String.format("%s.%s", issuerId, "FLIGHT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "FLIGHT_CLASS_SUFFIX"))));
 
     // Generic passes
     objectsToAdd.put(
         "genericObjects",
-        Arrays.asList(
-            new GenericObject()
-                .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
+            List.of(
+                    new GenericObject()
+                            .setId(String.format("%s.%s", issuerId, "GENERIC_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GENERIC_CLASS_SUFFIX"))));
 
     // Gift cards
     objectsToAdd.put(
         "giftCardObjects",
-        Arrays.asList(
-            new GiftCardObject()
-                .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
+            List.of(
+                    new GiftCardObject()
+                            .setId(String.format("%s.%s", issuerId, "GIFT_CARD_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "GIFT_CARD_CLASS_SUFFIX"))));
 
     // Loyalty cards
     objectsToAdd.put(
         "loyaltyObjects",
-        Arrays.asList(
-            new LoyaltyObject()
-                .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
+            List.of(
+                    new LoyaltyObject()
+                            .setId(String.format("%s.%s", issuerId, "LOYALTY_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "LOYALTY_CLASS_SUFFIX"))));
 
     // Offers
     objectsToAdd.put(
         "offerObjects",
-        Arrays.asList(
-            new OfferObject()
-                .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
+            List.of(
+                    new OfferObject()
+                            .setId(String.format("%s.%s", issuerId, "OFFER_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "OFFER_CLASS_SUFFIX"))));
 
     // Transit passes
     objectsToAdd.put(
         "transitObjects",
-        Arrays.asList(
-            new TransitObject()
-                .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
-                .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
+            List.of(
+                    new TransitObject()
+                            .setId(String.format("%s.%s", issuerId, "TRANSIT_OBJECT_SUFFIX"))
+                            .setClassId(String.format("%s.%s", issuerId, "TRANSIT_CLASS_SUFFIX"))));
 
     // Create the JWT as a HashMap object
     HashMap<String, Object> claims = new HashMap<String, Object>();
     claims.put("iss", ((ServiceAccountCredentials) credentials).getClientEmail());
     claims.put("aud", "google");
-    claims.put("origins", Arrays.asList("www.example.com"));
+    claims.put("origins", List.of("www.example.com"));
     claims.put("typ", "savetowallet");
     claims.put("payload", objectsToAdd);
 
@@ -854,7 +845,7 @@ public class DemoTransit {
     String token = JWT.create().withPayload(claims).sign(algorithm);
 
     System.out.println("Add to Google Wallet link");
-    System.out.println(String.format("https://pay.google.com/gp/v/save/%s", token));
+    System.out.printf("https://pay.google.com/gp/v/save/%s%n", token);
 
     return String.format("https://pay.google.com/gp/v/save/%s", token);
   }
@@ -866,9 +857,8 @@ public class DemoTransit {
    *
    * @param issuerId The issuer ID being used for this request.
    * @param classSuffix Developer-defined unique ID for this pass class.
-   * @throws IOException
    */
-  public void BatchCreateObjects(String issuerId, String classSuffix) throws IOException {
+  public void batchCreateObjects(String issuerId, String classSuffix) throws IOException {
     // Create the batch request client
     BatchRequest batch = service.batch(new HttpCredentialsAdapter(credentials));
 
@@ -912,11 +902,11 @@ public class DemoTransit {
                                       .setLanguage("en-US")
                                       .setValue("Hero image description"))))
               .setTextModulesData(
-                  Arrays.asList(
-                      new TextModuleData()
-                          .setHeader("Text module header")
-                          .setBody("Text module body")
-                          .setId("TEXT_MODULE_ID")))
+                      List.of(
+                              new TextModuleData()
+                                      .setHeader("Text module header")
+                                      .setBody("Text module body")
+                                      .setId("TEXT_MODULE_ID")))
               .setLinksModuleData(
                   new LinksModuleData()
                       .setUris(
@@ -930,27 +920,27 @@ public class DemoTransit {
                                   .setDescription("Link module tel description")
                                   .setId("LINK_MODULE_TEL_ID"))))
               .setImageModulesData(
-                  Arrays.asList(
-                      new ImageModuleData()
-                          .setMainImage(
-                              new Image()
-                                  .setSourceUri(
-                                      new ImageUri()
-                                          .setUri(
-                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
-                                  .setContentDescription(
-                                      new LocalizedString()
-                                          .setDefaultValue(
-                                              new TranslatedString()
-                                                  .setLanguage("en-US")
-                                                  .setValue("Image module description"))))
-                          .setId("IMAGE_MODULE_ID")))
+                      List.of(
+                              new ImageModuleData()
+                                      .setMainImage(
+                                              new Image()
+                                                      .setSourceUri(
+                                                              new ImageUri()
+                                                                      .setUri(
+                                                                              "http://farm4.staticflickr.com/3738/12440799783_3dc3c20606_b.jpg"))
+                                                      .setContentDescription(
+                                                              new LocalizedString()
+                                                                      .setDefaultValue(
+                                                                              new TranslatedString()
+                                                                                      .setLanguage("en-US")
+                                                                                      .setValue("Image module description"))))
+                                      .setId("IMAGE_MODULE_ID")))
               .setBarcode(new Barcode().setType("QR_CODE").setValue("QR code value"))
               .setLocations(
-                  Arrays.asList(
-                      new LatLongPoint()
-                          .setLatitude(37.424015499999996)
-                          .setLongitude(-122.09259560000001)))
+                      List.of(
+                              new LatLongPoint()
+                                      .setLatitude(37.424015499999996)
+                                      .setLongitude(-122.09259560000001)))
               .setPassengerType("SINGLE_PASSENGER")
               .setPassengerNames("Passenger names")
               .setTripType("ONE_WAY")
